pyproject.toml
[project]
name = "bullet-journal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fpdf2>=2.8.5",
    "jupyter>=1.1.1",
    "pydantic>=2.12.5",
]
